Change file structure and put components in their own folders/files
