+++ src/Square/square.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import '../index.css';
+
+function Square(props) {
+    return (
+        <button
+            className={"square " + (props.winner ? "winner" : null)}
+            onClick={props.onSquareClick}
+        >
+            {props.value}
+        </button>
+    );
+}
+
+export { Square };
